style:: 파일명 수정 및 컴포넌트 정리 - NullTable, ShareBtn 등 파일 분리를 통해 코드 정리 - API Request/Response Type 파일과 프론트단 Type 분리 - TokenToRem util 폴더에서 lib/constants 폴더로 이동
