+++ panda.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from '@pandacss/dev'
 
 import { buttonRecipe, chipRecipe, inputRecipe, labelRecipe, modalCardRecipe } from './src/lib/recipes/index'
-import { tokenToRem } from './src/util/tokenToRem'
+import { tokenToRem } from './src/lib/constants/tokenToRem'
 
 export default defineConfig({
   // Whether to use css reset
