+++ src/types/friends.ts
@@ -1,0 +1,17 @@
+export interface FriendInterface {
+  friendshipId: number
+  userId: number
+  name: string
+  username: string
+  major: string
+  language: string
+}
+
+export interface FriendRequestInterface {
+  friendshipId: number
+  userId: number
+  name: string
+  username: string
+  major: string
+  language: string
+}
