+++ src/util/timetableUtil.ts
@@ -1,5 +1,4 @@
-import { TimetableInfo } from '@/api/types/timetable'
-import { CourseType, DayType, Semester, SemesterType } from '@/types/timetable'
+import { CourseType, DayType, Semester, SemesterType, TimetableInfo } from '@/types/timetable'
 
 const dayToNumber: { [key in DayType]: number } = { Mon: 1, Tue: 2, Wed: 3, Thu: 4, Fri: 5, Sat: 6, Sun: 7 }
